+++ front-end/src/pages/user/common/Header/components/StoreSearchForm.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Input from "../../Input/Input";
-import { SearchInputForm } from "./StoreSearchFormLayout";
+import Input from "components/common/Input/Input";
+import { SearchInputForm } from "./StoreSearchForm.styles";
 import useGetStores from "pages/user/StoresPage/hooks/useGetStores";
 import IconBtn from "components/common/Button/icon/IconBtn";
 import { CiSearch } from "react-icons/ci";
